Added VR instead of a story
