+++ packages/react-components/react-slider/stories/src/Slider/index.stories.tsx
@@ -11,7 +11,6 @@ export { Step } from './SliderStep.stories';
 export { MinMax } from './SliderMinMax.stories';
 export { Vertical } from './SliderVertical.stories';
 export { Disabled } from './SliderDisabled.stories';
-export { Customized } from './SliderCustomized.stories';
 
 export default {
   title: 'Components/Slider',
